Move rocksdb to remote_map dir. Compare with undefined instead of checking for truthiness. Rename maxPoolSize and put().
