+++ packages/core/test/workspace/local/remote_map/remote_map.test.ts
@@ -32,7 +32,7 @@ import {
 } from '../../../../src/local-workspace/remote_map/remote_map'
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const rocksdbImpl = require('../../../../src/local-workspace/rocksdb').default
+const rocksdbImpl = require('../../../../src/local-workspace/remote_map/rocksdb').default
 
 const { serialize, deserialize } = serialization
 const { awu } = collections.asynciterable
